+++ cvat-ui/src/components/cvat-app.tsx
@@ -75,6 +75,7 @@ import EmailConfirmationPage from './email-confirmation-pages/email-confirmed';
 import EmailVerificationSentPage from './email-confirmation-pages/email-verification-sent';
 import IncorrectEmailConfirmationPage from './email-confirmation-pages/incorrect-email-confirmation';
 import CreateModelPage from './create-model-page/create-model-page';
+import PremiumFeaturesModal from './premium-features-modal/premium-features-modal';
 
 interface CVATAppProps {
     loadFormats: () => void;
@@ -128,6 +129,10 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
     public componentDidMount(): void {
         const core = getCore();
         const { history, location } = this.props;
+        const {
+            HEALTH_CHECK_RETRIES, HEALTH_CHECK_PERIOD, HEALTH_CHECK_REQUEST_TIMEOUT, SERVER_UNAVAILABLE_COMPONENT,
+            RESET_NOTIFICATIONS_PATHS,
+        } = appConfig;
         // configure({ ignoreRepeatedEventsWhenKeyHeldDown: false });
 
         // Logger configuration
@@ -140,13 +145,20 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
         EventRecorder.initSave();
 
         customWaViewHit(location.pathname, location.search, location.hash);
-        history.listen((_location) => {
-            customWaViewHit(_location.pathname, _location.search, _location.hash);
+        history.listen((newLocation) => {
+            customWaViewHit(newLocation.pathname, newLocation.search, newLocation.hash);
+            const { location: prevLocation } = this.props;
+            const shouldResetNotifications = RESET_NOTIFICATIONS_PATHS.from.some(
+                (pathname) => prevLocation.pathname === pathname,
+            );
+            const pathExcluded = shouldResetNotifications && RESET_NOTIFICATIONS_PATHS.exclude.some(
+                (pathname) => newLocation.pathname.includes(pathname),
+            );
+            if (shouldResetNotifications && !pathExcluded) {
+                this.resetNotifications();
+            }
         });
 
-        const {
-            HEALTH_CHECK_RETRIES, HEALTH_CHECK_PERIOD, HEALTH_CHECK_REQUEST_TIMEOUT, SERVER_UNAVAILABLE_COMPONENT,
-        } = appConfig;
         core.server.healthCheck(
             HEALTH_CHECK_RETRIES,
             HEALTH_CHECK_PERIOD,
@@ -360,6 +372,14 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
         }
     }
 
+    private resetNotifications(): void {
+        const { resetErrors, resetMessages } = this.props;
+
+        notification.destroy();
+        resetErrors();
+        resetMessages();
+    }
+
     // Where you go depends on your URL
     public render(): JSX.Element {
         const {
@@ -378,6 +398,8 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
             location,
             isModelPluginActive,
         } = this.props;
+
+        const { CVAT_BILLING_URL } = appConfig;
 
         const { healthIinitialized, backendIsHealthy } = this.state;
 
@@ -474,6 +496,7 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
                                     <ExportBackupModal />
                                     <ImportDatasetModal />
                                     <ImportBackupModal />
+                                    { CVAT_BILLING_URL && <PremiumFeaturesModal /> }
                                     {/* eslint-disable-next-line */}
                                     <a id='downloadAnchor' target='_blank' style={{ display: 'none' }} download />
                                 </Layout.Content>